RIP my dreams of a buildless frontend
managing deps & compat is still too much with workers and modules. it's
soooo close which makes it even more frustrating. it could probably work
if I was willing to give up using a worker for the connection, or was
less rigid about module scope. at least now I can switch to typescript.
revisit in a year to see where the browsers are, but by then I imagine
I'll be too dependent on ts.